// src/Utils/enums.ts
export enum ActivityTypeID {
  ACTIVITY = 1,
  CLUB = 2,
  CONTEST = 3,
  INTERN = 4,
  RECRUIT = 5,
}
export enum BoardSlug {
  ACTIVITY = "activity1",
  CONTEST = "contest",
  CLUB = "activity",
  INTERN = "interndata",
}

export enum BestPostType {
  recent = "RECENT",
  lastWeek = "LAST_WEEK",
  lastMonth = "LAST_MONTH",
  replyCount = "REPLY_COUNT",
  voteCount = "VOTE_COUNT",
}